backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
aiosqlite>=0.20.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
httpx>=0.27.0
alembic>=1.13.1
